refactor(font): extract nested endpoints into separate controller classes
